Get rid of overlayToolColorHelper (BL-11433)
